por enquanto tudo funciona, faltaria só contar trocas

+++ bolha.h
@@ -1,2 +1,2 @@
-void bolha(int *vetor, int n, int *comp, int *troca);
+void bolha(int *vetor, int n, long int *comp, long int *troca);
 void chamaBolha(int *vetor, int n);
